+++ src/core/bus.rs
@@ -62,13 +62,7 @@ impl Bus {
 
     fn log_message(message: &Message) {
         match &message.reading.value {
-            Value::Blob(content_type, content) => info!(
-                "[{:?}] {}: {} ({} bytes)",
-                &message.type_,
-                &message.sensor.id,
-                content_type,
-                content.len()
-            ),
+            Value::Blob(content) => info!("[{:?}] {}: {} bytes", &message.type_, &message.sensor.id, content.len()),
             ref value => info!("[{:?}] {} = {:?}", &message.type_, &message.sensor.id, value),
         }
     }